backend/requirements.txt
# SUBSTITUA: backend/requirements.txt
# Este é o teu ficheiro + o que faltava

# FastAPI e servidor
fastapi
uvicorn[standard]

# Banco de dados MongoDB
motor
pymongo
dnspython # ⭐️ ADICIONADO: Necessário para MongoDB Atlas

# Autenticação e segurança
bcrypt
PyJWT
python-multipart

# Modelos e validação
pydantic
pydantic-settings

# Variáveis de ambiente
python-dotenv

# Email (SendGrid)
sendgrid

# HTTP requests
requests

# Timezone e datas
python-dateutil

# Pagamentos (Mercado Pago)
mercadopago # ⭐️ ADICIONADO

# NOVO: backend/requirements.txt

# ... (restante do teu ficheiro) ...

# Firebase Admin SDK para acesso ao Firestore
firebase-admin